requirements.txt
# Discord
discord.py[voice]>=2.3.0
PyNaCl>=1.5.0

# Audio Processing
yt-dlp>=2023.10.13

# Spotify
spotipy>=2.23.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0

# Database
aiosqlite>=0.19.0

# Logging
colorlog>=6.7.0
